Refactor receipt processing by creating ReceiptStepper and ReceiptLineItemsTable components for improved modularity and readability
