+++ client/src/ReceiptStepper.tsx
@@ -1,0 +1,79 @@
+import { Stepper, Step, StepLabel, StepContent, Box, Typography } from '@mui/material'
+import ErrorIcon from '@mui/icons-material/Error'
+import CheckCircleIcon from '@mui/icons-material/CheckCircle'
+import { ReceiptLineItemsTable } from './ReceiptLineItemsTable'
+import type { Receipt } from 'shared'
+
+interface Props {
+  steps: string[]
+  activeStep: number
+  stepErrors: boolean[]
+  stepSuccess: boolean[]
+  logs: string[]
+  analyzedReceipt: Receipt | null
+}
+
+export function ReceiptStepper({
+  steps,
+  activeStep,
+  stepErrors,
+  stepSuccess,
+  logs,
+  analyzedReceipt,
+}: Props) {
+  return (
+    <Stepper activeStep={activeStep} orientation="vertical">
+      {steps.map((label, index) => (
+        <Step key={label} expanded={index <= activeStep}>
+          <StepLabel
+            error={stepErrors[index]}
+            StepIconComponent={({ completed, error }) => {
+              if (error) {
+                return <ErrorIcon color="error" />
+              } else if (completed || stepSuccess[index]) {
+                return <CheckCircleIcon color="success" />
+              } else {
+                return <span>{index + 1}</span>
+              }
+            }}
+          >
+            {label}
+          </StepLabel>
+          <StepContent>
+            {/* Custom rendering for Analyze Receipt step with table */}
+            {index === 1 && analyzedReceipt && analyzedReceipt.lineItems && analyzedReceipt.lineItems.length > 0 ? (
+              <Box>
+                <Typography variant="body2" sx={{ 
+                    whiteSpace: 'pre-line',
+                    color: stepErrors[index] ? 'error.main' : stepSuccess[index] ? 'success.main' : 'text.secondary'
+                  }}>
+                  ✓ Receipt analyzed successfully:
+                  <br />• Merchant: {analyzedReceipt.merchant}
+                  <br />• Date: {analyzedReceipt.transactionDate}
+                  <br />• Memo: {analyzedReceipt.memo}
+                </Typography>
+                <ReceiptLineItemsTable
+                  lineItems={analyzedReceipt.lineItems}
+                  totalTaxes={analyzedReceipt.totalTaxes}
+                  totalAmount={analyzedReceipt.totalAmount}
+                />
+              </Box>
+            ) : (
+              (index <= activeStep || logs[index]) && (
+                <Typography 
+                  variant="body2" 
+                  sx={{ 
+                    whiteSpace: 'pre-line',
+                    color: stepErrors[index] ? 'error.main' : stepSuccess[index] ? 'success.main' : 'text.secondary'
+                  }}
+                >
+                  {logs[index] || (index === activeStep ? 'In progress...' : '')}
+                </Typography>
+              )
+            )}
+          </StepContent>
+        </Step>
+      ))}
+    </Stepper>
+  )
+}
