+++ client/src/utils/formatLineItems.tsx
@@ -1,0 +1,9 @@
+import type { Receipt } from 'shared';
+
+export function formatLineItems(lineItems: Receipt['lineItems'] = []) {
+  return lineItems
+    .map(
+      (item) => `  - ${item.productName}: $${item.lineItemTotalAmount.toFixed(2)} (${item.category})`
+    )
+    .join('\n');
+}
